requirements.txt
requests~=2.28.1
python-dotenv~=0.21.0
python-telegram-bot==11.1.0
redis~=4.5.4
python-dotenv~=0.21.0